Shotguns are finally being made
Far from finished but it will hopefully work.

+++ Rouge-LikeCubeWorld/Assets/Scripts/Guns/GunStats.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class GunStats : MonoBehaviour
 {
     public float Damage;
-    public float MagSize;
+    public int MagSize;
     public float Price;
     public float Recoil;
 }
